Move components, stores and utils outside of apps folder
