+++ app/[boardId]/[pageId]/index.tsx
@@ -1,21 +1,21 @@
-import MessageWindow from "@/app/components/MessageWindow"
-import Page from "@/app/components/Page/Page"
-import PageAddSheet from "@/app/components/Page/PageAddSheet"
-import PageNav from "@/app/components/Page/PageNav"
-import { useBoards, usePagesetActions } from "@/app/stores/boards"
-import { useDebounceTime, useMessageWindowLocation } from "@/app/stores/prefs"
-import { generateNewPage } from "@/app/utils/boards"
-import { DebounceContext, handleDebounce } from "@/app/utils/debounce"
-import { handleError } from "@/app/utils/error"
+import MessageWindow from "@/components/MessageWindow"
+import Page from "@/components/Page/Page"
+import PageAddSheet from "@/components/Page/PageAddSheet"
+import PageNav from "@/components/Page/PageNav"
+import { useBoards, usePagesetActions } from "@/stores/boards"
+import { useDebounceTime, useMessageWindowLocation } from "@/stores/prefs"
+import { generateNewPage } from "@/utils/boards"
+import { DebounceContext, handleDebounce } from "@/utils/debounce"
+import { handleError } from "@/utils/error"
 import {
   loadManifest,
   loadPage,
   saveBoardPage,
   saveManifest,
-} from "@/app/utils/file"
-import { removePath } from "@/app/utils/io"
-import { useTheme } from "@/app/utils/theme"
-import { BoardButton, BoardPage, TileImage } from "@/app/utils/types"
+} from "@/utils/file"
+import { removePath } from "@/utils/io"
+import { useTheme } from "@/utils/theme"
+import { BoardButton, BoardPage, TileImage } from "@/utils/types"
 import { TrueSheet } from "@lodev09/react-native-true-sheet"
 import { Stack, useLocalSearchParams, useRouter } from "expo-router"
 import { useCallback, useEffect, useRef, useState } from "react"
