+++ app/create.tsx
@@ -1,11 +1,11 @@
 import { useRouter } from "expo-router"
 import { ScrollView, StyleSheet, View } from "react-native"
-import PageAdd from "./components/Page/PageAdd"
-import { usePagesetActions } from "./stores/boards"
-import { generateNewBoard } from "./utils/boards"
-import { saveBoard } from "./utils/file"
-import { GAP, MAX_WIDTH, PADDING, useTheme } from "./utils/theme"
-import { uuid } from "./utils/uuid"
+import PageAdd from "../components/Page/PageAdd"
+import { usePagesetActions } from "../stores/boards"
+import { generateNewBoard } from "../utils/boards"
+import { saveBoard } from "../utils/file"
+import { GAP, MAX_WIDTH, PADDING, useTheme } from "../utils/theme"
+import { uuid } from "../utils/uuid"
 
 export default function Create() {
   const theme = useTheme()
