+++ components/Settings/Button.tsx
@@ -1,4 +1,4 @@
-import { ICON_SIZE, useTheme } from "@/app/utils/theme"
+import { ICON_SIZE, useTheme } from "@/utils/theme"
 import { Link } from "expo-router"
 import { Settings } from "lucide-react-native"
 import { Button } from "../Styled"
