+++ components/SymbolPicker.tsx
