+++ components/Tile/Tile.tsx
@@ -1,7 +1,7 @@
-import { useButtonView, useLabelLocation } from "@/app/stores/prefs"
-import { useDebounce } from "@/app/utils/debounce"
-import { PADDING, RADIUS, useTheme } from "@/app/utils/theme"
-import { BoardButton } from "@/app/utils/types"
+import { useButtonView, useLabelLocation } from "@/stores/prefs"
+import { useDebounce } from "@/utils/debounce"
+import { PADDING, RADIUS, useTheme } from "@/utils/theme"
+import { BoardButton } from "@/utils/types"
 import { AACButton, AACStyle } from "@willwade/aac-processors/browser"
 import { StyleSheet } from "react-native"
 import { Pressable } from "react-native-gesture-handler"
