+++ stores/middleware.ts
