+++ stores/prefs.ts
