+++ utils/symbols.ts
